BackendServices/ChatService/requirements.txt
altgraph==0.17.3
aniso8601==9.0.1
anyio==3.6.2
asgiref==3.5.2
daphne==4.0.0 
async-timeout==4.0.2
autopep8==2.0.0
build==0.9.0
certifi==2022.9.24
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
customtkinter==5.0.0
darkdetect==0.7.1
decorator==5.1.1
distlib==0.3.6
Django==4.1.3
djangorestframework==3.14.0
fastapi==0.88.0
filelock==3.8.0
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
future==0.18.2
geocoder==1.38.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
pep517==0.13.0
Pillow==9.3.0
platformdirs==2.5.4
pycodestyle==2.10.0
pydantic==1.10.2
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
pyperclip==1.8.2
pytz==2022.6
ratelim==0.1.6
redis==4.4.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
SQLAlchemy-serializer==1.4.1
sqlparse==0.4.3
django-redis==5.2.0 
starlette==0.22.0
tkintermapview==1.17
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.16.7
Werkzeug==2.2.2
